requirements.txt
chess == 1.11.2
Pillow == 11.2.1